Fix running `create-rwjblue-release-it-setup`
The previous implementation failed when `create-rwjblue-release-it-setup` was not installed globally, or when the current working directory was not the one of the blueprint (e.g. running from `npx`). While `execa` supports running locally installed binaries by their name, by default it looks for them based on whatever `process.cwd()` is. So we need to tell `execa` to look for the binary based on the blueprint's directory.

Fixes #11