Update broken test-app script overrides
The default app blueprint has migrated away from `npm-run-all` to `concurrently`, but we had this override still in place, which now refers to a missing dependency. The main reason we had the override was the `test:watch` script, which we might actually not need really? (it forces us to exclude it from `test`, so makes our override very brittle)